Fixed FS Unlink

+++ src/helpers/delete_image.js
@@ -2,8 +2,11 @@ const fs = require('fs')
 
 module.exports = {
   deleteImage: (imgLoc) => {
-    fs.unlink(imgLoc, (error) => {
-      error ? console.log('Image not found') : console.log('Image deleted')
-    })
+    console.log(`Is this? ${imgLoc}`)
+    if (fs.existsSync(imgLoc)) {
+      fs.unlink(imgLoc, (error) => {
+        error ? console.log('Image not found') : console.log('Image deleted')
+      })
+    }
   }
 }
